Added try Catch

+++ src/main/java/com/test/restapi/RestAPI/resources/MessageResource.java
@@ -24,7 +24,7 @@ public class MessageResource {
 	
 	@GET
 	@Path("/{messageId}")
-	public Message getMessage(@PathParam("messageId") Long messageId){
+	public Message getMessage(@PathParam("messageId") Integer messageId){
 		return messageServices.getMessage(messageId);
 	}
